% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPhp.R
\name{asPhP}
\alias{asPhP}
\title{Convert number to Philippine Peso}
\usage{
asPhP(x)
}
\arguments{
\item{x}{a numeric vector or matrix}
}
\value{
a character vector or matrix of formatted decimals as Philippine Peso
}
\description{
Converts a numeric into Philippine Peso
}
