% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPhp.R
\name{asPhPFull}
\alias{asPhPFull}
\title{Convert number to Philippine Peso with Words}
\usage{
asPhPFull(x)
}
\arguments{
\item{x}{a number}
}
\value{
a character of formatted decimals as Philippine Peso
  with words
}
\description{
Converts a number into Philippine Peso with words
}
