% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center the table cell horizontally}
\usage{
center(x, toBold = TRUE, cols = 1, pos = "c")
}
\arguments{
\item{x}{a character vector}

\item{toBold}{logical value whether bold face fonts will be applied}

\item{cols}{integer value indicating the number of columns spanned}

\item{pos}{horizontal alignment of in the multicolumn macro}
}
\description{
Center the cell inside the table using the multicolumn macro
}
