% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Convert to percent}
\usage{
percent(x, latex = TRUE, digits = 2)
}
\arguments{
\item{x}{a number}

\item{latex}{logical value indicating if output is in latex}

\item{digits}{number of decimal digits to be used}
}
\value{
a character string
}
\description{
Converts a numeric to a percent format
}
