% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.R
\name{pluralize}
\alias{pluralize}
\title{Pluralize a word}
\usage{
pluralize(word, n)
}
\arguments{
\item{word}{a word to be pluralized}

\item{n}{a number describing the quantity of the word}
}
\value{
the word in plural form if the number is more than one
}
\description{
Converts a word to its plural form
}
